refactor: organise functions in separate files and directories (#3)
Move functions from index.ts to separate files and directories.
