+++ src/holidays/getBankHolidays.ts
@@ -1,0 +1,6 @@
+import type { HolidayList } from "../types";
+import { filterHolidays } from "../utils/filters";
+
+export function getBankHolidays(year: number): HolidayList {
+  return filterHolidays(year, (holiday) => holiday.bankHoliday);
+}
