+++ src/holidays/getChristmas.ts
@@ -1,0 +1,3 @@
+export function getChristmas(year: number) {
+  return new Date(year, 11, 25);
+}
