+++ src/holidays/getGoodFriday.ts
@@ -1,0 +1,6 @@
+import { addDays } from "date-fns";
+import { getEaster } from "./getEaster";
+
+export function getGoodFriday(year: number) {
+  return addDays(getEaster(year), -2);
+}
