+++ src/holidays/getHalloween.ts
@@ -1,0 +1,3 @@
+export function getHalloween(year: number) {
+  return new Date(year, 9, 31);
+}
