+++ src/holidays/getIndependenceDay.ts
@@ -1,0 +1,3 @@
+export function getIndependenceDay(year: number) {
+  return new Date(year, 6, 4);
+}
