+++ src/holidays/getJuneteenth.ts
@@ -1,0 +1,3 @@
+export function getJuneteenth(year: number) {
+  return new Date(year, 5, 19);
+}
