+++ src/holidays/getMemorialDay.ts
@@ -1,0 +1,5 @@
+import { getLastOfMonth } from "../utils/getters";
+
+export function getMemorialDay(year: number) {
+  return getLastOfMonth(new Date(year, 4), 1);
+}
