+++ src/holidays/getNewYearsDay.ts
@@ -1,0 +1,3 @@
+export function getNewYearsDay(year: number) {
+  return new Date(year, 0, 1);
+}
