+++ src/holidays/getNewYearsEve.ts
@@ -1,0 +1,5 @@
+import { lastDayOfMonth } from "date-fns";
+
+export function getNewYearsEve(year: number) {
+  return lastDayOfMonth(new Date(year, 11));
+}
