+++ src/holidays/getValentinesDay.ts
@@ -1,0 +1,3 @@
+export function getValentinesDay(year: number) {
+  return new Date(year, 1, 14);
+}
