+++ src/holidays/getVeteransDay.ts
@@ -1,0 +1,3 @@
+export function getVeteransDay(year: number) {
+  return new Date(year, 10, 11);
+}
